Create component limitDate to calendars
